requirements.txt
google-generativeai
networkx
pydantic
python-dotenv
requests
feedparser
fastapi
uvicorn
